requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
requests==2.32.3
beautifulsoup4==4.12.3
duckduckgo_search==7.5.2
google-generativeai==0.8.4
pydantic==2.10.6
python-dotenv==1.0.1
httpx>=0.27.0
openai>=1.0.0
